worker/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
httpx==0.25.2
together==1.3.3
python-dotenv==1.0.0